docs(time-series-chart-meta): rewrite stories in new format (#558)
