--- src/components/TimeseriesChartMeta/stories/customInterval.md
@@ -1,23 +1,0 @@
-## Custom Update Interval
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { TimeseriesChartMeta } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  return (
-    <TimeseriesChartMeta
-      timeseriesId={8681821313339919}
-      liveUpdate={true}
-      updateIntervalMillis={1000}
-    />
-  );
-  
-}
-```
