--- src/components/TimeseriesChartMeta/stories/predefinedPeriod.md
@@ -1,22 +1,0 @@
-## Predefined Period
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { TimeseriesChartMeta } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  return (
-    <TimeseriesChartMeta
-      timeseriesId={8681821313339919}
-      defaultTimePeriod="lastMonth"
-    />
-  );
-  
-}
-```
